(BSR) chore(archi): move FavoriteList inside Favorites folder
